Make bootloader key better protected
I have accidentally pressed the key combination for the bootloader
several times.  Make it harder to do this by mistake.  A positive effect
of this is also that it frees up one of the nicer keys on the bottom
layer.  I make it an additional MINUS key to spare my pinky some stress.